Lower swappiness even further.
Our platform default is now 10 which is in line with recommendations
"out there" (Red Hat amongst others) as a good setting to maximize
process usage of memory (instead of swapping out) instead of caches.

If users want caches to be used they need to tame their processes first.
From a platform perspective we really do not want to swap out processes
unnecessarily and rather drop VFS.