Work around failing password file substitutions
After refactoring the apg password generation code in f9cc724, nodes
tried to substitue the derivation created by fclib.servicePassword from
Hydra despite the fact that the derivation was marked with
`preferLocalBuild = true`. Hydra refused to deliver a substitution (410
Gone) and nixos-rebuild failed.

Now an arbitrary token is added to the generator drv. We can
individualize the password derivation now by including the local host
name in the token for example.

Case 107295