Fix firewall reload (again)
The fix provided in d8881bfce33 was not sufficient to regereate the
saved firewall rules in case just the system config changed but not the
config in /etc/local/firewall. This led to incorrect firewall rules
during bootstrap among others.

Now we save a checksum in /var/lib/firewall/startScript to rule out this
kind of situation. Note that the former `find` code is not necessary
anymore since the contents of /etc/local/firewall gets included into the
start script.