Wait for working nginx before requesting LE certificate, test
FCIO:

This is a patch for the upstream PR https://github.com/NixOS/nixpkgs/pull/60219.
It improved the situation but there were still failures because the
challenge was sent to the webserver before it was accepting requests for
the new domain. Maybe using another LE client would fix this, too.
Should be discussed with upstream.